+++ CHANGELOG.md
@@ -1,2 +1,7 @@
+* Fix pip 10 compatibility (use-python plugin 1.0.2)
+* Update default mkdocs 17.2 -> 17.3
+* Update default mkdocs-material 2.2.1 -> 2.7.2
+* Update default pymdown-extensions 4.6 -> 4.9.2
+
 ### 1.0.0 (2017-12-30)
 * Initial release